refactor and deleted dev code in this branch
