--- src/app/components/PlayButton/PlayButton.tsx
@@ -1,10 +1,0 @@
-const PlayButton = () => {
-    return (
-        <a href="https://futuregames.itch.io/yahwehs-creation" 
-        target="_blank" 
-        className="w-120 h-40 bg-[url('/images/playfreebtn.png')] bg-no-repeat bg-center bg-contain cursor-pointer"
-        aria-label="Play Yahweh's Creation"></a>
-    )
-}
-
-export default PlayButton;
